BAH-4687|Refactor. Extract BaseFileUpload base class
